examples/onto_interview/requirements.txt
openai>=1.0
streamlit>=1.30
pyyaml>=6.0
